requirements.txt
dotenv==0.9.9
exceptiongroup==1.3.1
iniconfig==2.3.0
mysql-connector==2.2.9
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
tomli==2.4.0
typing_extensions==4.15.0
uv==0.10.9
